+++ packages/components-js/projects/jsdom-polyfill/tests/unit/helper/index.ts
@@ -10,6 +10,7 @@ export const WHITELISTED_TAG_NAMES = [
   'p-multi-select-option',
   'p-optgroup',
   'p-select-option',
+  'p-radio-group-option',
   'p-stepper-horizontal-item',
   'p-tabs-item',
   'p-segmented-control-item',
@@ -138,6 +139,11 @@ export const tagNameMarkup: Record<TagNameRelevant, string> = {
     <p-segmented-control-item value="1">Item 1</p-segmented-control-item>
     <p-segmented-control-item value="2">Item 2</p-segmented-control-item>
   </p-segmented-control>`,
+  'p-radio-group': `<p-radio-group label="Some label" name="name" value="a">
+    <p-radio-group-option label="Option A" value="a"></p-radio-group-option>
+    <p-radio-group-option label="Option A" value="b"></p-radio-group-option>
+    <p-radio-group-option label="Option A" value="c"></p-radio-group-option>
+  </p-radio-group>`,
   'p-select': `<p-select name="name">
     <p-select-option value="a">Option A</p-select-option>
     <p-select-option value="b">Option B</p-select-option>