+++ packages/components/src/components.d.ts
@@ -53,15 +53,17 @@ import { MarqueAriaAttribute, MarqueTarget, MarqueVariant } from "./components/m
 import { MarqueSize } from "./components/marque/marque-size";
 import { ModalAriaAttribute, ModalBackdrop, ModalMotionHiddenEndEventDetail, ModalMotionVisibleEndEventDetail } from "./components/modal/modal-utils";
 import { ModelSignatureColor, ModelSignatureFetchPriority, ModelSignatureModel, ModelSignatureSize } from "./components/model-signature/model-signature-utils";
-import { MultiSelectDropdownDirection, MultiSelectState, MultiSelectUpdateEventDetail } from "./components/multi-select/multi-select/multi-select-utils";
+import { MultiSelectChangeEventDetail, MultiSelectDropdownDirection, MultiSelectState, MultiSelectUpdateEventDetail } from "./components/multi-select/multi-select/multi-select-utils";
 import { PaginationInternationalization, PaginationMaxNumberOfPageLinks, PaginationUpdateEventDetail } from "./components/pagination/pagination-utils";
-import { PinCodeLength, PinCodeState, PinCodeType, PinCodeUpdateEventDetail } from "./components/pin-code/pin-code-utils";
+import { PinCodeChangeEventDetail, PinCodeLength, PinCodeState, PinCodeType, PinCodeUpdateEventDetail } from "./components/pin-code/pin-code-utils";
 import { PopoverAriaAttribute, PopoverDirection } from "./components/popover/popover-utils";
 import { RadioButtonWrapperState } from "./components/radio-button-wrapper/radio-button-wrapper-utils";
+import { RadioGroupChangeEventDetail, RadioGroupDirection, RadioGroupState } from "./components/radio-group/radio-group/radio-group-utils";
+import { RadioGroupOptionBlurEventDetail } from "./components/radio-group/radio-group-option/radio-group-option-utils";
 import { ScrollerAlignScrollIndicator, ScrollerAriaAttribute, ScrollerGradientColor, ScrollerGradientColorScheme, ScrollerScrollIndicatorPosition, ScrollerScrollToPosition } from "./components/scroller/scroller-utils";
-import { SegmentedControlBackgroundColor, SegmentedControlColumns, SegmentedControlUpdateEventDetail } from "./components/segmented-control/segmented-control/segmented-control-utils";
+import { SegmentedControlBackgroundColor, SegmentedControlChangeEventDetail, SegmentedControlColumns, SegmentedControlUpdateEventDetail } from "./components/segmented-control/segmented-control/segmented-control-utils";
 import { SegmentedControlItemAriaAttribute, SegmentedControlItemIcon } from "./components/segmented-control/segmented-control-item/segmented-control-item-utils";
-import { SelectDropdownDirection, SelectState, SelectUpdateEventDetail } from "./components/select/select/select-utils";
+import { SelectChangeEventDetail, SelectDropdownDirection, SelectState, SelectUpdateEventDetail } from "./components/select/select/select-utils";
 import { SelectWrapperDropdownDirection, SelectWrapperState } from "./components/select-wrapper/select-wrapper/select-wrapper-utils";
 import { SheetAriaAttribute, SheetMotionHiddenEndEventDetail, SheetMotionVisibleEndEventDetail } from "./components/sheet/sheet-utils";
 import { SpinnerAriaAttribute, SpinnerSize } from "./components/spinner/spinner-utils";
@@ -129,15 +131,17 @@ export { MarqueAriaAttribute, MarqueTarget, MarqueVariant } from "./components/m
 export { MarqueSize } from "./components/marque/marque-size";
 export { ModalAriaAttribute, ModalBackdrop, ModalMotionHiddenEndEventDetail, ModalMotionVisibleEndEventDetail } from "./components/modal/modal-utils";
 export { ModelSignatureColor, ModelSignatureFetchPriority, ModelSignatureModel, ModelSignatureSize } from "./components/model-signature/model-signature-utils";
-export { MultiSelectDropdownDirection, MultiSelectState, MultiSelectUpdateEventDetail } from "./components/multi-select/multi-select/multi-select-utils";
+export { MultiSelectChangeEventDetail, MultiSelectDropdownDirection, MultiSelectState, MultiSelectUpdateEventDetail } from "./components/multi-select/multi-select/multi-select-utils";
 export { PaginationInternationalization, PaginationMaxNumberOfPageLinks, PaginationUpdateEventDetail } from "./components/pagination/pagination-utils";
-export { PinCodeLength, PinCodeState, PinCodeType, PinCodeUpdateEventDetail } from "./components/pin-code/pin-code-utils";
+export { PinCodeChangeEventDetail, PinCodeLength, PinCodeState, PinCodeType, PinCodeUpdateEventDetail } from "./components/pin-code/pin-code-utils";
 export { PopoverAriaAttribute, PopoverDirection } from "./components/popover/popover-utils";
 export { RadioButtonWrapperState } from "./components/radio-button-wrapper/radio-button-wrapper-utils";
+export { RadioGroupChangeEventDetail, RadioGroupDirection, RadioGroupState } from "./components/radio-group/radio-group/radio-group-utils";
+export { RadioGroupOptionBlurEventDetail } from "./components/radio-group/radio-group-option/radio-group-option-utils";
 export { ScrollerAlignScrollIndicator, ScrollerAriaAttribute, ScrollerGradientColor, ScrollerGradientColorScheme, ScrollerScrollIndicatorPosition, ScrollerScrollToPosition } from "./components/scroller/scroller-utils";
-export { SegmentedControlBackgroundColor, SegmentedControlColumns, SegmentedControlUpdateEventDetail } from "./components/segmented-control/segmented-control/segmented-control-utils";
+export { SegmentedControlBackgroundColor, SegmentedControlChangeEventDetail, SegmentedControlColumns, SegmentedControlUpdateEventDetail } from "./components/segmented-control/segmented-control/segmented-control-utils";
 export { SegmentedControlItemAriaAttribute, SegmentedControlItemIcon } from "./components/segmented-control/segmented-control-item/segmented-control-item-utils";
-export { SelectDropdownDirection, SelectState, SelectUpdateEventDetail } from "./components/select/select/select-utils";
+export { SelectChangeEventDetail, SelectDropdownDirection, SelectState, SelectUpdateEventDetail } from "./components/select/select/select-utils";
 export { SelectWrapperDropdownDirection, SelectWrapperState } from "./components/select-wrapper/select-wrapper/select-wrapper-utils";
 export { SheetAriaAttribute, SheetMotionHiddenEndEventDetail, SheetMotionVisibleEndEventDetail } from "./components/sheet/sheet-utils";
 export { SpinnerAriaAttribute, SpinnerSize } from "./components/spinner/spinner-utils";
@@ -2398,6 +2402,9 @@ export namespace Components {
          */
         "theme"?: Theme;
     }
+    /**
+     * @deprecated since v3.30.0, will be removed with next major release. Please use `p-radio-group` instead.
+     */
     interface PRadioButtonWrapper {
         /**
           * Show or hide label. For better accessibility it's recommended to show the label.
@@ -2423,6 +2430,82 @@ export namespace Components {
           * Adapts the color depending on the theme.
          */
         "theme"?: Theme;
+    }
+    interface PRadioGroup {
+        /**
+          * A boolean value that, if present, renders the radio group as a compact version.
+         */
+        "compact"?: boolean;
+        /**
+          * Supplementary text providing more context or explanation for the radio group.
+         */
+        "description"?: string;
+        /**
+          * Defines the direction of the main and cross axis. The default is 'column' showing options vertically stacked. You always need to provide a base value when using breakpoints.
+         */
+        "direction"?: BreakpointCustomizable<RadioGroupDirection>;
+        /**
+          * A boolean value that, if present, makes the radio group unusable and unclickable. The value will not be submitted with the form.
+         */
+        "disabled"?: boolean;
+        /**
+          * Specifies the id of the <form> element that the radio group belongs to (useful if the radio group is not a direct descendant of the form).
+         */
+        "form"?: string;
+        /**
+          * Controls the visibility of the label.
+         */
+        "hideLabel"?: BreakpointCustomizable<boolean>;
+        /**
+          * Text content for a user-facing label.
+         */
+        "label"?: string;
+        /**
+          * @experimental Shows a loading indicator.
+         */
+        "loading"?: boolean;
+        /**
+          * Dynamic feedback text for validation or status.
+         */
+        "message"?: string;
+        /**
+          * The name of the group of radio buttons, used when submitting the form data.
+         */
+        "name": string;
+        /**
+          * A boolean value that specifies a selection must be made from the group before the form can be submitted.
+         */
+        "required"?: boolean;
+        /**
+          * Indicates the validation or overall status of the radio group component.
+         */
+        "state"?: RadioGroupState;
+        /**
+          * Controls the visual appearance of the component.
+         */
+        "theme"?: Theme;
+        /**
+          * The default value for the radio-group.
+         */
+        "value"?: string;
+    }
+    interface PRadioGroupOption {
+        /**
+          * A boolean value that, if present, makes the radio group option unusable and unclickable.
+         */
+        "disabled"?: boolean;
+        /**
+          * Text content for a user-facing label.
+         */
+        "label"?: string;
+        /**
+          * @experimental Shows a loading indicator.
+         */
+        "loading"?: boolean;
+        /**
+          * The value for the input.
+         */
+        "value"?: string;
     }
     interface PScroller {
         /**
@@ -3255,6 +3338,14 @@ export interface PPinCodeCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLPPinCodeElement;
 }
+export interface PRadioGroupCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLPRadioGroupElement;
+}
+export interface PRadioGroupOptionCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLPRadioGroupOptionElement;
+}
 export interface PSegmentedControlCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLPSegmentedControlElement;
@@ -3885,6 +3976,8 @@ declare global {
         new (): HTMLPModelSignatureElement;
     };
     interface HTMLPMultiSelectElementEventMap {
+        "blur": void;
+        "change": MultiSelectChangeEventDetail;
         "update": MultiSelectUpdateEventDetail;
     }
     /**
@@ -3938,6 +4031,8 @@ declare global {
         new (): HTMLPPaginationElement;
     };
     interface HTMLPPinCodeElementEventMap {
+        "blur": void;
+        "change": PinCodeChangeEventDetail;
         "update": PinCodeUpdateEventDetail;
     }
     /**
@@ -3963,11 +4058,48 @@ declare global {
         prototype: HTMLPPopoverElement;
         new (): HTMLPPopoverElement;
     };
+    /**
+     * @deprecated since v3.30.0, will be removed with next major release. Please use `p-radio-group` instead.
+     */
     interface HTMLPRadioButtonWrapperElement extends Components.PRadioButtonWrapper, HTMLStencilElement {
     }
     var HTMLPRadioButtonWrapperElement: {
         prototype: HTMLPRadioButtonWrapperElement;
         new (): HTMLPRadioButtonWrapperElement;
+    };
+    interface HTMLPRadioGroupElementEventMap {
+        "change": RadioGroupChangeEventDetail;
+    }
+    interface HTMLPRadioGroupElement extends Components.PRadioGroup, HTMLStencilElement {
+        addEventListener<K extends keyof HTMLPRadioGroupElementEventMap>(type: K, listener: (this: HTMLPRadioGroupElement, ev: PRadioGroupCustomEvent<HTMLPRadioGroupElementEventMap[K]>) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | AddEventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLPRadioGroupElementEventMap>(type: K, listener: (this: HTMLPRadioGroupElement, ev: PRadioGroupCustomEvent<HTMLPRadioGroupElementEventMap[K]>) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | EventListenerOptions): void;
+    }
+    var HTMLPRadioGroupElement: {
+        prototype: HTMLPRadioGroupElement;
+        new (): HTMLPRadioGroupElement;
+    };
+    interface HTMLPRadioGroupOptionElementEventMap {
+        "blur": RadioGroupOptionBlurEventDetail;
+    }
+    interface HTMLPRadioGroupOptionElement extends Components.PRadioGroupOption, HTMLStencilElement {
+        addEventListener<K extends keyof HTMLPRadioGroupOptionElementEventMap>(type: K, listener: (this: HTMLPRadioGroupOptionElement, ev: PRadioGroupOptionCustomEvent<HTMLPRadioGroupOptionElementEventMap[K]>) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | AddEventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLPRadioGroupOptionElementEventMap>(type: K, listener: (this: HTMLPRadioGroupOptionElement, ev: PRadioGroupOptionCustomEvent<HTMLPRadioGroupOptionElementEventMap[K]>) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | EventListenerOptions): void;
+    }
+    var HTMLPRadioGroupOptionElement: {
+        prototype: HTMLPRadioGroupOptionElement;
+        new (): HTMLPRadioGroupOptionElement;
     };
     interface HTMLPScrollerElement extends Components.PScroller, HTMLStencilElement {
     }
@@ -3977,6 +4109,8 @@ declare global {
     };
     interface HTMLPSegmentedControlElementEventMap {
         "segmentedControlChange": SegmentedControlUpdateEventDetail;
+        "blur": void;
+        "change": SegmentedControlChangeEventDetail;
         "update": SegmentedControlUpdateEventDetail;
     }
     /**
@@ -4003,6 +4137,8 @@ declare global {
         new (): HTMLPSegmentedControlItemElement;
     };
     interface HTMLPSelectElementEventMap {
+        "blur": void;
+        "change": SelectChangeEventDetail;
         "update": SelectUpdateEventDetail;
     }
     /**
@@ -4380,6 +4516,8 @@ declare global {
         "p-pin-code": HTMLPPinCodeElement;
         "p-popover": HTMLPPopoverElement;
         "p-radio-button-wrapper": HTMLPRadioButtonWrapperElement;
+        "p-radio-group": HTMLPRadioGroupElement;
+        "p-radio-group-option": HTMLPRadioGroupOptionElement;
         "p-scroller": HTMLPScrollerElement;
         "p-segmented-control": HTMLPSegmentedControlElement;
         "p-segmented-control-item": HTMLPSegmentedControlItemElement;
@@ -6684,7 +6822,15 @@ declare namespace LocalJSX {
          */
         "name"?: string;
         /**
+          * Emitted when the multi-select has lost focus.
+         */
+        "onBlur"?: (event: PMultiSelectCustomEvent<void>) => void;
+        /**
           * Emitted when the selection is changed.
+         */
+        "onChange"?: (event: PMultiSelectCustomEvent<MultiSelectChangeEventDetail>) => void;
+        /**
+          * @deprecated since v3.30.0, will be removed with next major release, use `change` event instead. Emitted when the selection is changed.
          */
         "onUpdate"?: (event: PMultiSelectCustomEvent<MultiSelectUpdateEventDetail>) => void;
         /**
@@ -6827,7 +6973,15 @@ declare namespace LocalJSX {
          */
         "name"?: string;
         /**
-          * Emitted when selected element changes.
+          * Emitted when the pin-code has lost focus.
+         */
+        "onBlur"?: (event: PPinCodeCustomEvent<void>) => void;
+        /**
+          * Emitted when the input is changed.
+         */
+        "onChange"?: (event: PPinCodeCustomEvent<PinCodeChangeEventDetail>) => void;
+        /**
+          * @deprecated since v3.30.0, will be removed with next major release, use `change` event instead. Emitted when the input is changed.
          */
         "onUpdate"?: (event: PPinCodeCustomEvent<PinCodeUpdateEventDetail>) => void;
         /**
@@ -6869,6 +7023,9 @@ declare namespace LocalJSX {
          */
         "theme"?: Theme;
     }
+    /**
+     * @deprecated since v3.30.0, will be removed with next major release. Please use `p-radio-group` instead.
+     */
     interface PRadioButtonWrapper {
         /**
           * Show or hide label. For better accessibility it's recommended to show the label.
@@ -6894,6 +7051,90 @@ declare namespace LocalJSX {
           * Adapts the color depending on the theme.
          */
         "theme"?: Theme;
+    }
+    interface PRadioGroup {
+        /**
+          * A boolean value that, if present, renders the radio group as a compact version.
+         */
+        "compact"?: boolean;
+        /**
+          * Supplementary text providing more context or explanation for the radio group.
+         */
+        "description"?: string;
+        /**
+          * Defines the direction of the main and cross axis. The default is 'column' showing options vertically stacked. You always need to provide a base value when using breakpoints.
+         */
+        "direction"?: BreakpointCustomizable<RadioGroupDirection>;
+        /**
+          * A boolean value that, if present, makes the radio group unusable and unclickable. The value will not be submitted with the form.
+         */
+        "disabled"?: boolean;
+        /**
+          * Specifies the id of the <form> element that the radio group belongs to (useful if the radio group is not a direct descendant of the form).
+         */
+        "form"?: string;
+        /**
+          * Controls the visibility of the label.
+         */
+        "hideLabel"?: BreakpointCustomizable<boolean>;
+        /**
+          * Text content for a user-facing label.
+         */
+        "label"?: string;
+        /**
+          * @experimental Shows a loading indicator.
+         */
+        "loading"?: boolean;
+        /**
+          * Dynamic feedback text for validation or status.
+         */
+        "message"?: string;
+        /**
+          * The name of the group of radio buttons, used when submitting the form data.
+         */
+        "name"?: string;
+        /**
+          * Emitted when the selected option is changed.
+         */
+        "onChange"?: (event: PRadioGroupCustomEvent<RadioGroupChangeEventDetail>) => void;
+        /**
+          * A boolean value that specifies a selection must be made from the group before the form can be submitted.
+         */
+        "required"?: boolean;
+        /**
+          * Indicates the validation or overall status of the radio group component.
+         */
+        "state"?: RadioGroupState;
+        /**
+          * Controls the visual appearance of the component.
+         */
+        "theme"?: Theme;
+        /**
+          * The default value for the radio-group.
+         */
+        "value"?: string;
+    }
+    interface PRadioGroupOption {
+        /**
+          * A boolean value that, if present, makes the radio group option unusable and unclickable.
+         */
+        "disabled"?: boolean;
+        /**
+          * Text content for a user-facing label.
+         */
+        "label"?: string;
+        /**
+          * @experimental Shows a loading indicator.
+         */
+        "loading"?: boolean;
+        /**
+          * Emitted when the radio input has lost focus.
+         */
+        "onBlur"?: (event: PRadioGroupOptionCustomEvent<RadioGroupOptionBlurEventDetail>) => void;
+        /**
+          * The value for the input.
+         */
+        "value"?: string;
     }
     interface PScroller {
         /**
@@ -6954,11 +7195,19 @@ declare namespace LocalJSX {
          */
         "name"?: string;
         /**
+          * Emitted when the segmented-control has lost focus.
+         */
+        "onBlur"?: (event: PSegmentedControlCustomEvent<void>) => void;
+        /**
+          * Emitted when the selection is changed.
+         */
+        "onChange"?: (event: PSegmentedControlCustomEvent<SegmentedControlChangeEventDetail>) => void;
+        /**
           * @deprecated since v3.0.0, will be removed with next major release, use `update` event instead. Emitted when selected element changes.
          */
         "onSegmentedControlChange"?: (event: PSegmentedControlCustomEvent<SegmentedControlUpdateEventDetail>) => void;
         /**
-          * Emitted when selected element changes.
+          * @deprecated since v3.30.0, will be removed with next major release, use `change` event instead. Emitted when selected element changes.
          */
         "onUpdate"?: (event: PSegmentedControlCustomEvent<SegmentedControlUpdateEventDetail>) => void;
         /**
@@ -7041,7 +7290,15 @@ declare namespace LocalJSX {
          */
         "name"?: string;
         /**
+          * Emitted when the select has lost focus.
+         */
+        "onBlur"?: (event: PSelectCustomEvent<void>) => void;
+        /**
           * Emitted when the selection is changed.
+         */
+        "onChange"?: (event: PSelectCustomEvent<SelectChangeEventDetail>) => void;
+        /**
+          * @deprecated since v3.30.0, will be removed with next major release, use `change` event instead. Emitted when the selection is changed.
          */
         "onUpdate"?: (event: PSelectCustomEvent<SelectUpdateEventDetail>) => void;
         /**
@@ -7775,6 +8032,8 @@ declare namespace LocalJSX {
         "p-pin-code": PPinCode;
         "p-popover": PPopover;
         "p-radio-button-wrapper": PRadioButtonWrapper;
+        "p-radio-group": PRadioGroup;
+        "p-radio-group-option": PRadioGroupOption;
         "p-scroller": PScroller;
         "p-segmented-control": PSegmentedControl;
         "p-segmented-control-item": PSegmentedControlItem;
@@ -7939,7 +8198,12 @@ declare module "@stencil/core" {
              */
             "p-pin-code": LocalJSX.PPinCode & JSXBase.HTMLAttributes<HTMLPPinCodeElement>;
             "p-popover": LocalJSX.PPopover & JSXBase.HTMLAttributes<HTMLPPopoverElement>;
+            /**
+             * @deprecated since v3.30.0, will be removed with next major release. Please use `p-radio-group` instead.
+             */
             "p-radio-button-wrapper": LocalJSX.PRadioButtonWrapper & JSXBase.HTMLAttributes<HTMLPRadioButtonWrapperElement>;
+            "p-radio-group": LocalJSX.PRadioGroup & JSXBase.HTMLAttributes<HTMLPRadioGroupElement>;
+            "p-radio-group-option": LocalJSX.PRadioGroupOption & JSXBase.HTMLAttributes<HTMLPRadioGroupOptionElement>;
             "p-scroller": LocalJSX.PScroller & JSXBase.HTMLAttributes<HTMLPScrollerElement>;
             /**
              * @controlled { "props": ["value"], "event": "update", "isInternallyMutated": true }